Refactored code - moved code to new GymStatusTab.tsx
