+++ frontend/src/components/ConfirmationModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from '../Button'
+import Button from './Button'
 import styled from 'styled-components'
 
 const OverlayContainer = styled.div<{ overlayBackgroundColor: string }>`
