+++ frontend/src/components/Gym/GymStatusTab.tsx
@@ -1,0 +1,118 @@
+import React, { ReactElement, useEffect, useState } from 'react'
+import styled from 'styled-components'
+import GymStatus from '../GymStatus'
+import GymKeyWith from '../GymKeyWith'
+import GymConfirmationModal from './GymConfirmationModal'
+import { useDispatch, useSelector } from 'react-redux'
+import { RootState } from '../../store/types'
+import { getGymStatus } from '../../store/gym/action'
+import { getUserDetail } from '../../store/social/action'
+import { ButtonStates, ButtonTypes } from '../../store/gym/types'
+import ButtonComponent from '../Button'
+
+const GymContainer = styled.div`
+  padding: 50px 0;
+`
+
+const ButtonContainer = styled.div`
+  min-height: 150px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+`
+
+function GymStatusTab() {
+  const dispatch = useDispatch()
+  const { gymStatus } = useSelector((state: RootState) => state.gym)
+  const { name } = useSelector((state: RootState) => state.social)
+  useEffect(() => {
+    dispatch(getGymStatus())
+    dispatch(getUserDetail())
+    // dispatch(getGymHistory())
+  }, [])
+
+  const [isModalOpen, setModalOpen] = useState<boolean>(false)
+  const [modalState, setModalState] = useState<ButtonTypes>()
+
+  const buttonMap = new Map<ButtonTypes, string>([
+    ['keyWithMe', 'Key With Me'],
+    ['returnKey', 'Return Key'],
+    ['closeGym', 'Close Gym'],
+    ['openGym', 'Open Gym'],
+  ])
+
+  const gymButtonStates = new Map<string, ButtonStates>([
+    ['gymOpen', { keyWithMe: false, returnKey: false, toggleGym: true }],
+    ['gymClosed', { keyWithMe: false, returnKey: true, toggleGym: true }],
+    ['keyWithOthers', { keyWithMe: true, returnKey: false, toggleGym: false }],
+  ])
+
+  function getButtonStates(): ButtonStates {
+    if (!gymStatus.gymIsOpen) {
+      if (gymStatus.keyHolder.displayName == name) {
+        return gymButtonStates.get('gymClosed')!
+      }
+      return gymButtonStates.get('keyWithOthers')!
+    }
+    //Runs this if gym is open
+    if (gymStatus.keyHolder.displayName == name) {
+      return gymButtonStates.get('gymOpen')!
+    }
+    return gymButtonStates.get('keyWithOthers')!
+  }
+
+  function getButtonText(key: string): string {
+    if (key === 'toggleGym') {
+      return gymStatus.gymIsOpen ? buttonMap.get('closeGym')! : buttonMap.get('openGym')!
+    }
+    return buttonMap.get(key)!
+  }
+
+  function handleModalState(key: string): void {
+    setModalOpen(true)
+    if (key === 'toggleGym') {
+      if (gymStatus.gymIsOpen) {
+        setModalState('closeGym')
+        return
+      }
+      setModalState('openGym')
+      return
+    }
+    setModalState(key)
+  }
+
+  function renderConfirmationModal(): ReactElement {
+    return (
+      <GymConfirmationModal
+        gymIsOpen={gymStatus.gymIsOpen}
+        name={name}
+        setModalOpen={setModalOpen}
+        isModalOpen={isModalOpen}
+        modalState={modalState}
+      />
+    )
+  }
+
+  function renderButton() {
+    const buttonState: ButtonStates = getButtonStates()
+    return Object.keys(buttonState).map((key: string, index: number) => (
+      <ButtonComponent
+        key={index}
+        state={buttonState[key] ? 'primary' : 'secondary'}
+        text={getButtonText(key)}
+        onClick={() => handleModalState(key)}
+      />
+    ))
+  }
+
+  return (
+    <GymContainer>
+      {renderConfirmationModal()}
+      <GymStatus isOpen={gymStatus.gymIsOpen} />
+      <GymKeyWith name={gymStatus.keyHolder.displayName} handle={gymStatus.keyHolder.telegramHandle} />
+      <ButtonContainer>{renderButton()}</ButtonContainer>
+    </GymContainer>
+  )
+}
+export default GymStatusTab
